tratamento de erro/problema com put e delete

+++ backend/src/controllers/DoadorController.js
@@ -2,15 +2,29 @@ const Usuario = require('../models/Usuario');
 
 module.exports = {
     async adicionaDoador(req, res) {
-        const doc = req.body;
+        const { id, nome, email, celular, classe } = req.body;
 
-        const doadorExiste = await Usuario.findOne({ id: doc.id });
+        if(!id) {
+            return res.status(400).json({ error: "Entrada invalida: id do usuario nao pode ser vazio ou nulo." });
+        }
+        else if(!nome) {
+            return res.status(400).json({ error: "Entrada invalida: nome nao pode ser vazio ou nulo." });
+        }
+        else if(!email) {
+            return res.status(400).json({ error: "Entrada invalida: email nao pode ser vazio ou nulo." });
+        }
+        else if(!celular) {
+            return res.status(400).json({ error: "Entrada invalida: celular nao pode ser vazio ou nulo." });
+        }
+        else if(!classe) {
+            return res.status(400).json({ error: "Entrada invalida: classe nao pode ser vazia ou nula." });
+        }
+
+        const doadorExiste = await Usuario.findOne({ id: id });
 
         if(doadorExiste) {
             return res.json(doadorExiste);
         }
-
-        const { id, nome, email, celular, classe } = req.body;
 
         const doador = await Usuario.create({
             id,
@@ -25,6 +39,10 @@ module.exports = {
 
     async pesquisaUsuarioPorId(req, res) {
         const doc = req.params.id;
+
+        if(!id) {
+            return res.status(400).json({ error: "Entrada invalida: id do usuario nao pode ser vazio ou nulo." });
+        }
 
         const doadorExiste = await Usuario.findOne({ id: doc });
 
@@ -41,10 +59,14 @@ module.exports = {
     async pesquisaUsuarioPorNome(req, res) {
         const nome = req.params.nome;
 
+        if(!nome) {
+            return res.status(400).json({ error: "Entrada invalida: nome nao pode ser vazio ou nulo." });
+        }
+
         const doadorExiste = await Usuario.findOne({ nome: nome });
 
         if(!doadorExiste) {
-            return res.status(400).json({ error: 'Usuário não encontrado: nome'});
+            return res.status(400).json({ error: `Usuário não encontrado: ${nome}`});
         }
 
         var status = (doadorExiste.classe == "pessoa fisica" ? "doador" : "receptor");
@@ -53,8 +75,12 @@ module.exports = {
        
     },
 
-    async atualizaUsuario(req, res) {
+    async atualizaUsuario(req, res) {   //problema para pegar vazio na rota
         const doc = req.params.id;
+
+        if(!doc) {
+            return res.status(400).json({ error: "Entrada invalida: id do usuario nao pode ser vazio ou nulo." });
+        }
 
         const update = req.body;
 
@@ -64,18 +90,22 @@ module.exports = {
             });
             return res.json(atualizaDoador);
         } catch (err) {
-            return res.status(400).json({ error: 'Usuário não encontrado: id' });
+            return res.status(400).json({ error: `Usuário não encontrado: ${doc}` });
         }
     },
 
-    async removeUsuario(req, res) {
+    async removeUsuario(req, res) {   //problema para pegar vazio na rota
         const doc = req.params.id;
+
+        if(!doc) {
+            return res.status(400).json({ error: "Entrada invalida: id do usuario nao pode ser vazio ou nulo." });
+        }
 
         try {
             const removeUsuario = await Usuario.findOneAndDelete({ id: doc });
             return res.json(removeUsuario);
         } catch (err) {
-            return res.status(400).json({ error: 'Usuário não encontrado: id' });
+            return res.status(400).json({ error: `Usuário não encontrado: ${doc}` });
         }
     }
 };
